Add files via upload
INTRODUCTION

Airbnb is a company that offers visitors, travellers, and tourists an internet platform for lodging. It offers a wide range of rooms and homestays that are hosted by individuals who have registered as Airbnb hosts. These rental properties include apartments, homes, and boats in various neighbourhoods all over the world. Property rates vary by location and region, as well as neighbourhood and room availability.

New York City is one of the most popular tourist destinations, attracting visitors from all over the world. Properties that cater to various needs based on pricing.


OBJECTIVE

For the past few months, Airbnb has seen a major decline in revenue. Now that the restrictions have started lifting and people have started to travel more, Airbnb wants to make sure that it is full prepared for this change and what are the following parameters required to increase the revenue



ABOUT THE DATASET

The dataset has been downloaded from upgrad student portal

1.	The data set depicts the listing activity and analytics in New York City. It contains information about the hosts, their regional availability, as well as reviews and ratings. It features 16 columns with information such as the host id, host name, latitude and longitude, reviews, room type, 365-day availability, and so on.